Add child types to content.
i.e. Annotation.SUB_ITEMS -> Content::getContent
     Annoration.SUB_ITEM_SUMMARIES -> Content::getSubItemSummaries
     Annotation.SERIES -> Content::getSeries
Note: the latter has forced a backwards incompatible change in the
Content::getSeries signature, as a episode has a series singluar, but a
brand has series plural. So now Content::getSeries returns a list, which
will be a singleton if the type is episode.